# packages/api-client/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "Object": {
        "properties": {},
        "type": "object"
      },
...
      "PaginatedEventsResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EventResponse"
            },
            "type": "array"
          },
          "page": {
            "description": "1-based page number.",
            "example": 1,
            "type": "number"
          },
          "pageSize": {
            "example": 25,
            "type": "number"
          },
          "total": {
            "description": "Total events matching the filter.",
            "example": 7,
            "type": "number"
          }
        },
        "required": ["items", "page", "pageSize", "total"],
        "type": "object"
      },
...
      "ProblemDetails": {
        "properties": {
          "code": {
            "example": "VALIDATION_ERROR",
            "type": "string"
          },
          "detail": {
            "example": "One or more request values are invalid.",
            "type": "string"
          },
          "errors": {
            "description": "Field-level details when the failure is a validation error.",
            "type": "object"
          },
          "instance": {
            "example": "/api/v1/auth/login",
            "type": "string"
          },
          "requestId": {
            "example": "018f2c9e-1d3a-7b21-9c44-2f1a6b5d0e77",
            "type": "string"
          },
          "status": {
            "example": 400,
            "type": "number"
          },
          "title": {
            "example": "Validation Failed",
            "type": "string"
          },
          "type": {
            "example": "https://api.event-platform.local/problems/validation_error",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code",
          "requestId"
        ],
        "type": "object"
      },
...
      "UpdateEventDto": {
        "properties": {
          "description": {
            "maxLength": 2000,
            "minLength": 1,
            "nullable": true,
            "type": "string"
          },
          "endAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "eventType": {
            "enum": [
              "FILM_PREMIERE",
              "CONCERT",
              "ALBUM_RELEASE",
              "PRODUCT_LAUNCH",
              "CORPORATE_EVENT",
              "PROMOTIONAL_EVENT",
              "OTHER"
            ],
            "type": "string"
          },
          "location": {
            "maxLength": 200,
            "minLength": 1,
            "nullable": true,
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string"
          },
          "organizerName": {
            "maxLength": 200,
            "minLength": 1,
            "nullable": true,
            "type": "string"
          },
          "startAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "cookie": {
        "description": "Short-lived access token cookie",
        "in": "cookie",
        "name": "access_token",
        "type": "apiKey"
      },
      "csrf-token": {
        "description": "CSRF token for state-changing authenticated requests",
        "in": "header",
        "name": "x-csrf-token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Versioned API for the Event and Creative Operations Management Platform",
    "title": "Event and Creative Operations Platform API",
    "version": "1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/events": {
      "get": {
        "operationId": "Events_list_v1",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "enum": [
                "PLANNING",
                "READY",
                "IN_PROGRESS",
                "COMPLETED",
                "CANCELLED"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "eventType",
            "required": false,
            "schema": {
              "enum": [
                "FILM_PREMIERE",
                "CONCERT",
                "ALBUM_RELEASE",
                "PRODUCT_LAUNCH",
                "CORPORATE_EVENT",
                "PROMOTIONAL_EVENT",
                "OTHER"
              ],
              "type": "string"
            }
          },
          {
            "description": "Restrict to events whose connected workspace this user manages.",
            "in": "query",
            "name": "managerId",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "Case-insensitive match against the event name.",
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "maxLength": 200,
              "type": "string"
            }
          },
          {
            "description": "Only events that start at or after this UTC instant.",
            "in": "query",
            "name": "startingAfter",
            "required": false,
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "description": "Only events that start at or before this UTC instant.",
            "in": "query",
            "name": "startingBefore",
            "required": false,
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ],
              "default": 1,
              "minimum": 1
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ],
              "default": 25,
              "maximum": 100,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedEventsResponse"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Missing the required permission, or CSRF token invalid"
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "List events the caller may see, with filters",
        "tags": ["Event management"]
      },
      "post": {
        "operationId": "Events_create_v1",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Missing the required permission, or CSRF token invalid"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The requested manager user does not exist"
          }
        },
        "security": [
          {
            "access_token": []
          },
          {
            "csrf-token": []
          }
        ],
        "summary": "Create an event and its connected workspace",
        "tags": ["Event management"]
      }
    },
    "/api/v1/events/{id}": {
      "delete": {
        "operationId": "Events_remove_v1",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The event was removed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Missing the required permission, or CSRF token invalid"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Event not found"
          }
        },
        "security": [
          {
            "access_token": []
          },
          {
            "csrf-token": []
          }
        ],
        "summary": "Remove an event and its connected workspace",
        "tags": ["Event management"]
      },
      "get": {
        "operationId": "Events_get_v1",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Missing the required permission, or CSRF token invalid"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Event not found"
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Get one event with its connected workspace overview",
        "tags": ["Event management"]
      },
      "patch": {
        "operationId": "Events_update_v1",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Missing the required permission, or CSRF token invalid"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Event not found"
          }
        },
        "security": [
          {
            "access_token": []
          },
          {
            "csrf-token": []
          }
        ],
        "summary": "Update event details (not status, manager, teams, or budget)",
        "tags": ["Event management"]
      }
    },
    "/api/v1/events/{id}/budget": {
      "get": {
        "operationId": "Events_getBudget_v1",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventBudgetResponse"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Missing the required permission, or CSRF token invalid"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Event not found"
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Read the event budget (sensitive)",
        "tags": ["Event management"]
      },
      "put": {
        "operationId": "Events_setBudget_v1",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetEventBudgetDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventBudgetResponse"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Missing the required permission, or CSRF token invalid"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Event not found"
          }
        },
        "security": [
          {
            "access_token": []
          },
          {
            "csrf-token": []
          }
        ],
        "summary": "Set or clear the event budget (sensitive)",
        "tags": ["Event management"]
      }
    },
    "/api/v1/events/{id}/manager": {
      "put": {
        "operationId": "Events_setManager_v1",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignEventManagerDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Missing the required permission, or CSRF token invalid"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The event or the requested manager user does not exist"
          }
        },
        "security": [
          {
            "access_token": []
          },
          {
            "csrf-token": []
          }
        ],
        "summary": "Set, change, or clear the event manager",
        "tags": ["Event management"]
      }
    },
    "/api/v1/events/{id}/teams/{teamId}": {
      "delete": {
        "operationId": "Events_unassignTeam_v1",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "teamId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Missing the required permission, or CSRF token invalid"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Event not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "That team is not assigned to this event"
          }
        },
        "security": [
          {
            "access_token": []
          },
          {
            "csrf-token": []
          }
        ],
        "summary": "Unassign a team from this event",
        "tags": ["Event management"]
      },
      "put": {
        "operationId": "Events_assignTeam_v1",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "teamId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Missing the required permission, or CSRF token invalid"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The event or the team does not exist"
          }
        },
        "security": [
          {
            "access_token": []
          },
          {
            "csrf-token": []
          }
        ],
        "summary": "Assign a team to this event (idempotent)",
        "tags": ["Event management"]
      }
    },
    "/api/v1/events/{id}/transition": {
      "post": {
        "operationId": "Events_transition_v1",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionEventDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Missing the required permission, or CSRF token invalid"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Event not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "That transition is not allowed from the current state"
          }
        },
        "security": [
          {
            "access_token": []
          },
          {
            "csrf-token": []
          }
        ],
        "summary": "Move an event to another lifecycle state",
        "tags": ["Event management"]
      }
    },
...
  },
  "servers": [],
...
}